requirements.txt
pandas>=2.0
numpy>=1.24
statsmodels>=0.14
scikit-learn>=1.3
matplotlib>=3.7
PyYAML>=6.0
jupyter>=1.0
nbconvert>=7.10
